Add test for wrong bucketId order in CommitLogCheckpoint.create
Added to make it clear that the order of the keys in the map is important. A
better solution would be, of course, to make the order unimportant.

This can be done by using ImmutableSortedMap either as input or with copyOf
the input. But the create function is only used once, so it doesn't seem worth the effort - as long as we make sure to throw an error if the order is wrong.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=164515959