Fix incorrect use of TestLogHandler that causes tests to be flaky
The following test failed repeatedly and consistently during testing of
[] - part of the work to upgrade mockito.
//third_party/java_src/gtld/javatests/google/registry/flows:EppControllerTest

The test failure was not caused by the mockito change, although it's possible
that the mockito change did change something that made it more likely to occur.
The failure was caused because the test added a TestLogHandler to a Logger in
order to intercept a LogRecord but did not maintain a strong reference to the
Logger. So sometimes it was GCed before the EppController class obtained its
own reference to it. That meant that the Logger that produced the LogRecord
was not the same as the Logger to which the TestLogHandler was attached and so
the LogRecord was never intercepted and the test failed.

This change keeps a strong reference and also removes the handler after the
test has finished.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=171824566