Count EppResources that don't have an "old enough" revision
DeleteOldCommitLogs uses the revisions of EPP resources to know which commit logs are still in use.

This means that if the revision list of some EPP resources is corrupt in any way, we might delete commit logs that are still needed.

Here we add a basic sanity check for the revision list: we check it has at least one revision older than the "threshold", where the "threshold" is the oldest date we need to be able to recreate using the commit logs.

If we find an error, we log it (and increment a counter). There's nothing we can to do fix it thought.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=172780418