Change nonfunctional markdown-style anchors to HTML
The RDAP help page has anchors, so that we can then change the internal RDAP help topics to point to specific parts of the documentation. Unfortunately, GitHub does not appear to support the style of anchor we were using. See the current situation:

https://github.com/google/nomulus/blob/master/docs/rdap.md

Experimentation in GitHub confirms that switching to embedded HTML <a> tags solves the problem, and Critique preview shows it working correctly as well, with automatically generated table of contents at the left. If there's another way to fix this problem, I'm happy to try it.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=173405974