Increase number of frontend/backend instances on prod/sandbox to 100
The higher the number the better for serious launches. These used to be 100
but had been detuned because instances weren't dying correctly when no longer
needed, thus contributing to higher costs than necessary. That problem was
fixed when we migrated to the Java 8 runtime, however, so there's no reason
not to use the higher number.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=184742738