Reduce OAuth token cache time to 30min
It seems that even though the token is supposed to be valid for 60min, in
practice it expires before that. Reducing caching time to 30min solves the
problem (at least as far as I can tell). This should not increase too much load
as we are only calling the API twice an hour instead of once.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=186830395